ubifs: mount as -osync
Summary:
According to the ubifs documentation[1], it is by default an async file
system and operates different from jffs2 in the event of sudden power
drops.  Since our code was not explicitly written to handle async file
system operations (there is a sum total of 1 `fsync` call in the entire
repository), it is safer to revert back to jffs2 file-system semantics.

Change the mount option to be `-osync`.

1. http://www.linux-mtd.infradead.org/doc/ubifs.html#L_sync_semantics

Test Plan:
Enabled in QEMU and observed sync mount.

```
root@bmc-oob:/# mount | grep ubifs
/dev/ubi0_0 on /mnt/data type ubifs (rw,sync,relatime,compr=zstd,assert=read-only,ubi=0,vol=0)
```

Reviewed By: amithash

fbshipit-source-id: acc51d4685