temprary remove attestation api for fby3 cit test
Summary:
the existing rest-api CIT test will not work for the new api: `/api/attestation`
After discuss with rpanchal1996 of api usage, I'll develop a new CIT suite to test this. So for now we'll skip this API in the legacy CIT rest test.

Test Plan: normal regression CIT test

Reviewed By: peterdelevoryas

fbshipit-source-id: e4a013a905