rackmon2: Add abstraction for a poll thread
Summary:
Allows for a polling thread to be woken up when stopped.
This should replace the previous suboptimal method
where a signal handler should wait all of the 2min for
each thread to exit.

The PollThread class nicely abstracts away the worker
poll thread and all the user needs to do is operate on
the start/stop methods along with the constructors which
pass in the class member method, object and the polling
interval

Test Plan: Create test executable using the method.

Reviewed By: garnermic

fbshipit-source-id: ff603a27f4