rackmond2: Handle large buffers over the UNIX socket
Summary:
The rackmond UNIX socket protocol is very minimal, has a
2 byte length field followed by the buffer of len bytes.
Unfortunately, this means our data cannot be larger than
0xffff bytes. But this is incorrect, a data command can
stream a JSON file (currently 89k and slated to be larger
in the future). V1 managed this by using the protocol only
for the RAW command and just have the data stream without
the len header over to the target. This is inconsistent at best.

Work around this by splitting the data into 0xffff byte chunks
and streaming each across instead of making the protocol
inconsistent or breaking backwards compatibility by
changing the length field to be 4 bytes.

Test Plan: Test rackmoncli data and ensure that we get the full data back.

Reviewed By: zhdaniel12

fbshipit-source-id: d28e39278c