rackmond2: Add capability to format the register value/value-history
Summary:
This finally brings all the register format descriptor to use
and allows the user to request for a string representation of
a individual register value or a register store.

Currently it returns a string in a pretty human readable format.
This is not designed to be consumed by automation and hence we
should not write scripts depending on this. If such a thing is
required, we might need a new interface which returns a variant
of types (std::variant<int,float,std::vector<uint8_t>,std::string, etc).

Test Plan:
Test
```
rackmoncli data --format
<snip>
"  <0000>                        MFG_MODEL :700-014671-0000  ",
<snip>
```

Reviewed By: zhdaniel12

fbshipit-source-id: cd3cde3756