rackmond2: Add syslog capability of logs
Summary:
Send over logs to syslog instead of printing to stdout/stderr.
The abstraction for the std::ostream is in log.hpp. This allows
one to just replace this header when porting to a new platform
if they do not prefer syslog or want the dependency on syslog.

Test Plan:
Build and load pkg on wedge400 and ensure that
/var/log/messages contain the new logs.

Reviewed By: zhdaniel12

fbshipit-source-id: 952bee2c4f