rackmon2: Move device creation into its own virtual method
Summary:
This accomplishes two things. One, its better organization
to have a make_device() factory, seconds we can now mock
the make_device() factory to return a mocked device to
allow this to be unit-testable.

Profile store was converted from a std::stringstream& to
a std::ostream&. This allows rackmon to pass in any
ostream. Heck, it could pass in std::cout or std::syslog!

Test Plan: All unit-tests pass

Reviewed By: tao-ren

fbshipit-source-id: 6b7244505b