meta-aspeed: tune: add redirect for newer Yocto
Summary:
The latest Yocto moved the Aspeed 'tune' files.  Create an
indirect so that if the current (old) location doesn't exist we will
automatically redirect to the new location.  This allows systems with
older distros and those with newer distros to co-exist by using the
older tuning location.

Test Plan: Built Northdome with Honister.

Reviewed By: doranand

fbshipit-source-id: fb3020927c