rackmon2: (refactor) Move address interpretation to DESIGN.md
Summary:
Having the information on how we interpret the address of a modbus
device in regmap.hpp was confusing. It had nothing to do with
the rest of the header! It in fact is better to know if we are
reading the JSON files. Hence moved it to DESIGN.md.

Also add comments to describe the data structures.

Test Plan: None, refactor.

Reviewed By: tao-ren

fbshipit-source-id: 711e7cfa2c