rackmon2: Add capability to specify constant literals
Summary:
Allow users to be able to specify a raw message as a constant
literal. Example
```
Msg msg = 0x1234abcd_M;
```
This hopefully would be a lot more easier to not get lost
in tests.

We also onboard modbus_test.cpp and modbus_cmds_test.cpp
as a demonstration on how simple some of the tests could be.

Test Plan: All unit tests pass

Reviewed By: garnermic

fbshipit-source-id: 218063f1cd