rackmon2: Fix invalid memory access due to static corruption
Summary:
This was a bug which the unit-tests (future diff) exposed.
Having a static iterator point to the begin() of an object
not yet constructed at the time of static creation is
terrible undefined behavior.

*it ended up pointing to some random object. We were potentally
stomping all especially when incrementing it.

Fix it by making it an object member.

Test Plan: All unit tests pass

Reviewed By: garnermic

fbshipit-source-id: fb91b7476e