circle-ci: cit: Add --skip-qemu-setup option
Summary:
This is useful for local testing, and fixing stuff. The intended way to use this is to start QEMU yourself, with whatever image you want to use for testing, and then run `cit_test_wrapper.py` with `--skip-qemu-setup` so that it just scp's and runs tests. This also avoids waiting for QEMU to boot the image everytime you rerun this.

Test Plan:

    python3 tools/circle-ci/cit_test_wrapper.py fby3 -t tests.fby3.test_cfg_util.CfgUtilTest.test_cfg_util_set --skip-qemu-setup
    start qemu cit test on platform: fby3
    scp tests to BMC complete..
    qemu flag set..
    start CIT run in QEMU
    tests.fby3.test_cfg_util.CfgUtilTest.test_cfg_util_set: pass
    CIT test complete, PASS: 1, FAIL: 0, TOTAL: 1