sensors.py: Replace URI to master with URI to rev
Summary:
This accompanies D33795541 and is not meant to be merged until that change is accepted and pushed, so that I can get a chance to change the symlink to gvfs.

I've both updated the recipe here, and now that D33795541 is merged to the source_mirrors repository, I've updated the symlink to gvfs in this commit too. I also updated Dunfell using the new lf-obmc method prior to this commit.

Test Plan:
Verify that everything builds successfully once the source mirrors are updated.

Also I submitted these changes as a pull request to test it in CircleCI, and all the image building jobs succeeded (previously, some platforms like wedge100 weren't). https://app.circleci.com/pipelines/github/facebook/openbmc?branch=pull%2F172&filter=all

Reviewed By: deathowl

fbshipit-source-id: 3f7046abfa