tests2: Add QEMU denylist mechanism
Summary:
Right now, we use a decorator to specify if we should skip QEMU tests:

```
@unittest.skipIf(qemu_check(), "test env is QEMU, skipped")
class I2cTest(BaseI2cTest, unittest.TestCase):
    def load_golden_i2c_tree(self):
        self.i2c_tree = plat_i2c_tree
```

This mechanism works ok, but is difficult to maintain because often-times hundreds of tests will start failing, requiring hundreds of decorators to be added, and often requiring finding and overriding base class methods from `tests2/common/`. Right now we have 232 decorators doing this:

```
$ grep -r "skipIf" tests2 | wc -l
232
```

I would like to propose an alternative mechanism, a "qemu denylist", which specifies all the tests which should not be run in QEMU because they would fail.

This is also useful because if somebody wants to see if a test can be verified in QEMU, they can check the denylist and verify that it's _not_ disabled there.

Specifying this information as a denylist, rather than an allowlist, minimizes the file size.

Test Plan:
Making sure CircleCI and Sandcastle jobs pass.