circleci: Fix rest-api unit-test
Summary:
`rest-api` unit tests are failing on the Flake8 step in CircleCI (but not in Sandcastle), this refactors the CircleCI setup to allow Flake8 to pass.

Full context on this fix:

- `rest-api` unit tests use Flake8 to type check Python 3 type annotations.
- Flake8 uses Python 3 `multiprocessing`.
- `multiprocessing` uses POSIX semaphores.
- POSIX semaphores use `/dev/shm` on Linux ([man7](https://man7.org/linux/man-pages/man7/sem_overview.7.html))
- `/dev/shm` is a `tmpfs` virtual filesystem.
- So, Flake8 needs `/dev/shm`.
- Unit tests run inside the `qemux86-image` rootfs using `chroot`.
- The `qemux86-image` rootfs doesn't have `/dev/shm`.
- In Sandcastle, we manually create `.../qemux86-image-rootfs/dev/shm` and mount a 512MB `tmpfs` there.
- In CircleCI, we _cannot_ create new `tmpfs` mounts ([Forum thread](https://discuss.circleci.com/t/will-privileged-containers-sys-admin-and-dev-fuse-be-supported/35869)).
- So, `flake8` fails in CircleCI with "permission denied" while trying to mount `/dev/shm`.

The fix in this diff (#thanks @patrickw3 for suggesting [this](https://circleci.com/docs/2.0/executor-types/#ram-disks)) is to unpack the `qemux86-image` rootfs _within_ an existing `tmpfs`, in this case, `/mnt/ramdisk`. That way, `/dev/shm` is automatically a `tmpfs`.

I've also decoupled the `build-qemux86` and `unit-test` jobs in this diff as well, so that we can run `ptest-runner` as `root`, but still build the `qemux86-image` as a non-root user (which `bitbake` requires). This is so that we can move the `mount` command from the rest-api recipe to `ptest-runner`.

Lastly, to move the `/dev/shm` mounting step from the rest-api recipe to `ptest-runner`, I needed to refactor the shell command we use when entering the chroot to run a series of commands everytime we enter the chroot. Everytime we exit the chroot, the mount will disappear.

Test Plan:
Making sure CircleCI and Sandcastle unit-tests pass.