tests2: Replace qemu_denylist with a generic denylist file
Summary:
`cit_runner.py` checks for the existence of the file `/usr/local/bin/tests2/dummy_qemu` to determine if we're in QEMU. That, in turn, determines whether we use `qemu_denylist.txt` to filter out tests.

This turns out to not work well in Netcastle, because even in QEMU Netcastle runs, we run `cit_runner.py` outside of QEMU to list tests. We need a better way to specify when to use the denylist.

I think the simplest thing is to just make it an optional argument: in CircleCI and QEMU Netcastle runs, we can specify the denylist path in the invocation, and everywhere else we can just ignore the denylist argument.

Test Plan:
Watching CircleCI to make sure it doesn't break.