tests2: Add --tmpfs argument to ptest-runner
Summary:
I changed ptest-runner to unpack the rootfs in the working directory, and then forced CircleCI to cd to /mnt/ramdisk, so that the rootfs would get unpacked in a tmpfs. Instead, we could just specify the tmpfs directory as an optional argument for this situation, and keep the old behavior for everything else.

Test Plan:
Verifying CircleCI and Sandcastle ptest runs pass.