rackmon2: Remove illegal register address from monitor rotation
Summary:
It is not unusual for a modbus device PSU to not support
reading from certain registers. This is especially common
for registers added in a subsequent generation of the PSU.

This change works on the Modbus error framework to handle
and disable the register from monitoring if we get the
ILLEGAL_DATA_ADDRESS ModbusError exception.

Since a PSU could be unplugged and replaced with a newer
generation supporting the given address, when we re-activate
the PSU, we enable all registers.

Test Plan:
All unit tests pass.
Build and load on wedge400 and notice logging of registers
only once.

Reviewed By: zhdaniel12

fbshipit-source-id: d895dcf1f7