treewide: switch S location to default
Summary:
For a while we have been plagued with 'pseudo abort' issues
whenever rebuilding many of our packages.  After digging deep into the
problem (strace-ing a few bitbake executions) the cause ends up being
that Yocto doesn't provide good support for our "ship source with the
recipes" working model.  Many of our recipes set `S = "${WORKDIR}"`
because the default location for fetching `file://foo.cpp` is
`${WORKDIR}` but when we do that there is explicit code[1] in Yocto that
disables the pseudo-ignore for `${S}`.  This, in turn, causes a
pseudo-conflict between the source fetching (file://) and the
PACKAGE_DEBUG_SPLIT support, which is what causes the 'pseudo-abort'
anytime a single source file changes.

We have two possible solutions:
    1. Unset the `S = "${WORKDIR}"` directives.
    2. Set INHIBIT_PACKAGE_DEBUG_SPLIT=1.

Setting INHIBIT_PACKAGE_DEBUG_SPLIT would reduce the usability of GDB so
I've chosen to unset the `S = "${WORKDIR}"` directive.  This causes us
to use the default of something like `${WORKDIR}/${BPN}-${PV}`.  One
side-effect of this is that the default source location is ${WORKDIR},
so we need to use a `;subdir=${S}` directive on all of our `file://`
URIs.

Added a bbclass which defines a new variable `LOCAL_URI` to assist with
this `subdir` URI transformation, inject it into all recipes via the
common layer `INHERIT`, and update appropriate recipes to utilize it.
Also, in some cases we incorrectly used `${WORKDIR}` instead of `${S}`
in `do_install` steps.

All of this has been resolved in a sweeping change to the majority of
the recipes in the tree.

1. https://github.com/openbmc/openbmc/blob/9a2a1dade6b355dafe6e72cfb26732ca5e12587f/poky/meta/classes/base.bbclass#L436

Test Plan:
Built a few systems locally.  Let CIT build determine any
other breakages.

Reviewed By: amithash, zhdaniel12

fbshipit-source-id: 2b8cdbeff7