rackmon2: Remove support for deprecated string/print data
Summary:
After "value" data was added which is a structured interpretation
of the value stored in the registers, it formally removed
any value of having "print format". But since we did not have
CLI properly or usably print the "value" data, it was held
temoprarily.

Now that CLI implements a proper parsed/formatted display of
the value data, we can formally remove this functionality.

Test Plan:
build and all unit tests pass.
copy IPK to wedge400 and ensure it displays data correctly.

Reviewed By: garnermic

fbshipit-source-id: 8df3f5b144