rackmon2: Clean up req/resp messages upon command failure
Summary:
Modbus could have encoded/decoded the message or in other words
changed the length of the contents.
Req: We are guaranteed that encode is either non-stateful (Can be
repeated with the same initial len or wont change raw[] position if
some bytes are present (The only use of this is WriteMultiple)).
Resp: The only thing we want from this is the length of the
response expected, so in this case, we are good.

Test Plan:
Add unit test to explicitly test this regression.
Make sure the test fails without the fix for Modbus and
passes with the above fix.

Reviewed By: GoldenBug

fbshipit-source-id: 273e30e5441e181fe26a2830d2dfed6822f92cbf