libpldm: remove explicit IMAGE_INSTALL directives
Summary:
We only need explicit IMAGE_INSTALL directives for packages
which are not picked up via a different dependency.  Typically this is
the executables and daemons you want installed.  Packages which are
pure-libraries (like pldm) should not be explicitly added to the
IMAGE_INSTALL because they are already picked up by dependency
relationships.

Test Plan:
Run `bitbake <image> -gn` and check that pldm is still
included in the `pn-buildlist`.

Reviewed By: amithash

fbshipit-source-id: faa13e5342f75ef7d097deae75c23749b1b7d7c2