dhcpcd: add wrapper for dhclient emulation
Summary:
We have a whole bunch of initscripts and if-up hooks that call
into dhclient, but dhclient doesn't exist anymore starting with
Yocto-Honister (it has been replaced with dhcpcd).  Write a shell-script
which will emulate dhclient by calling dhcpcd with the correct
corresponding arguments.

Test Plan:
Booted Northdome in QEMU with Honister and now we get an IP
address.

Reviewed By: zhdaniel12

fbshipit-source-id: 501a7f192a52611a999281977cdfd4da7f49c47e