rackmon2: Fix large timeouts observed on wedge400
Summary:
The current write method goes in 4 steps.
1. disableRead
2. write request
3. wait for LSR assert.
4. enableRead
5. Read response
Normally, a response would arrive after #4 thus,
marking the transaction as successful.
But occationally (1/10 times) the response
arrives before #4 thus causing the device to
completely miss the response -- resulting in a timeout.

This change, replaces this with
1. Keeping read enabled throughout.
2. ignore LSR changes (Do not do #3).
This change completely eliminates timeouts on wedge400+ with
an FTDI USB device which do not need this sequencing.

But we would still need this functionality for wedge100 where we use
the native aspeed device (Pending a verification on wedge100. If it
is not needed even on that platform, that functionality can be
completely stripped.)

Test Plan:
On one SSH session, run
`modbuscmd -t 300 -x 7 b50300000001`
a thousand times and on a parallel SSH session run
`modbuscmd -t 300 -x 7 b60300000001`
a thousand times.
Each iteration ensure we get the correct response and after the stress
test ensure that the number of timeouts are moderate (<10).

Reviewed By: garnermic

fbshipit-source-id: 7846f2875cd32f402c766f5e2599bf4171ba41b3