rackmon2: Speed up modbuscmd (Legacy compatibility)
Summary:
Current legacy compatibility wraps the old command using
a shell script. This unfortunately adds a 100-200ms bump in
delay when executing.
This can add up, for 6 PSUs it can be a slow down of 1s.

This adds the support into rackmoncli itself and create
a link from modbuscmd to rackmoncli instead of a wrapper
script.

Test Plan:
Build, load on wedge400 and observe that the
median latency observed by tooling is within 2-3s.

Reviewed By: GoldenBug

fbshipit-source-id: 69c3c3603ab84bb9ff8c9c66e2f211016f3c5e6a