yocto: license mapping for older distros
Summary:
There is an incompatibility in the values of the LICENSE
variable between Honister/Kirkstone and Dunfell/Rocko.  In order to
provide compatibility across all the distributions we need to provide
a LICENSE mapping using the SPDXLICENSEMAP variable.

In order to accomplish this make the following changes:
    1. Add a mechanism to automatically import a
       `conf/distro/openbmc-yocto-<version>.inc`.
    2. Convert the previous `openbmc-pin-<version>` to leverage this
       automatic include instead.
    3. Delete a stray `openbmc-pin-zeus`.
    4. Add the necessary SPDXLICENSEMAP in order to support future
       Honister/Kirkstone recipe LICENSE updates.

Test Plan: Successfully built machines in CIT.

Reviewed By: doranand

fbshipit-source-id: 491f63a4bccf11dd51d3c1693edee8ebf5d0df02