dhcpcd: wait for MAC assign on NC-SI
Summary:
On NC-SI NICs we need to wait for the MAC to be assigned to the device,
which is indicated by the `addr_assign_type` going to `3` ("set using
dev_set_mac_address").  Add a hook to any machines using `mf-ncsi` to
add this wait for up to 30 seconds when using `dhcpcd`.  Also, add a hack
for QEMU so we skip the 30 second wait.

Test Plan:
Ran in QEMU on Northdome.  ODM tested on a test build
with Northdome on Honister (where dhcpcd is used) and provided the
basic `wait_mac_addr_ready` code.

Reviewed By: amithash

fbshipit-source-id: 1daf0d8a75091ff9dabaa247cf7a9a92ec5e7718