rackmon2: Fix failed probe on wedge100
Summary:
A couple of assumptions were wrong.

First and foremost, the IOCTL TIOCSERGETLSR is non-blocking.
Thus, we cannot assume that we do not need to do a second
iteration on it (The comment was wrong).
Hence, as observed on wedge100, we end up with a lot of
prints "Waited loops: 100". Change this to print only
when we hit the max iterations, not every time we have
more than 1 iteration.

Secondly, 100 iterations is very often exceeded. We would need
to bump this up to 10k to bring in some reliability.

Test Plan:
Probe and stress test (firmware upgrade usually does it)
using rackmon2 on wedge100.

Reviewed By: GoldenBug

fbshipit-source-id: 9dc03af9b576a2a283bd77809e2ab160c4613169