rackmon2: Execute raw command requests immediately
Summary:
The motivation of this change is to improve latency
of raw requests and hence lower the end-to-end latency
as observed by dynamo.

1. Make the UnixSock abstraction defer the thread creation
to the service rather than create a detached thread for
each connection. This gives the service freedom to either
execute in context or asynchronously.
2. Have the Rackmon service execute the "raw" command
synchronously and other long running commands asynchronously.

This brings the median latency observed by dynamo (per iteration)
from 3.5s to 2.5s.

Test Plan:
All unit tests pass.
Stress and performance tests on wedge100 and wedge400.

Reviewed By: GoldenBug

fbshipit-source-id: 0f04110a1aaf0b1c85c66168d2ce0f0ea53817e7