tpm: packagegroup: force non-allarch package (again)
Summary:
In D36644356, I attempted to force our TPM package group
to not be an 'allarch' package.  The code was correct but placed
in the wrong location per the upstream commit that adds the sanity
check.  Move it to the correct location in the packagegroup file.

This wasn't noticed in the previous test plan because it seems that
IPKs are not explicitly re-written unless forced to be.

Test Plan:
Execute the below...

```
$ bitbake packagegroup-openbmc-tpm -c cleanall
$ bitbake packagegroup-openbmc-tpm -ne | grep "^PACKAGE_ARCH="
PACKAGE_ARCH="armv7at2hf-vfp"
$ bitbake packagegroup-openbmc-tpm -c package_write_ipk
```

Reviewed By: amithash

fbshipit-source-id: 64828afa1d6a936db5eef7721149bd9551f7090d