halfdome: force DISTRO_CODENAME
Summary:
The manifold tools used to cache sources internally expects the
DISTRO_CODENAME to align with the name of the yocto branch we
are using for our source.  Halfdome is attempting to build with
the `lf-master` yocto branch, but the current default DISTRO_CODENAME
there is "langdale", which will be the next release of Yocto.

Force it to 'master' just like we do in `openbmc-fb-lf.conf` which
is used for Bletchley for the same reasons.

Test Plan:
Examine `bitbake halfdome-image -ne` output and observe:

```
 # $DISTRO_CODENAME [2 operations]
 #   set
 .../yocto/lf-master/meta-poky/conf/distro/poky.conf:4
 #     "langdale"
 #   set
 .../yocto/lf-master/../../meta-facebook/conf/distro/openbmc-fb-master.conf:5
 #     "master"
 # pre-expansion value:
 #   "master"
 DISTRO_CODENAME="master"
```

Reviewed By: peterdelevoryas

fbshipit-source-id: aef0a199c8885c9320f0e13f5f69b91bfa4bc4aa