churchillmono: adjust proprietary recipe inclusion
Summary:
In order for the Churchillmono machine to build in both internal
and OSS contexts, we cannot rely on the bblayers.conf.sample because
this will attempt to add the layer in the OSS context which doesn't
have it.

Add a new bbclass, fb-proprietary-recipes, which simply sets a variable
OVERRIDE of the same name.  The internal version of fb-only-network
will enable said bbclass and it can be enabled manually by users as
needed with `INHERIT:append = " fb-proprietary-recipes"`.

Add code in the `lf-obmc` mirror fetcher to include the proprietary
recipes as needed.

Change the churchillmono bblayers.conf and layer.conf to move from
the BBLAYERS method to the new BBFILES method, leveraging the
fb-proprietary-recipes override via:
`FB_PROPRIETARY_RECIPES:fb-proprietary-recipes = "..."` so that the
recipes from the proprietary tree are only included if the
fb-proprietary-recipes bbclass is utilized.

Test Plan:
Attempt to fetch the `cisco-wb-kmod` recipe in various
combinations of `fb-proprietary-recipes` and `fb-only-network` to
confirm that the recipe is found or not as appropriate.

Reviewed By: binhuang00

fbshipit-source-id: 46cd587f4590260e7f724f0d53124471274ed254