rackmon2: Add support for special handling for modbus devices
Summary:
This formally unveils support to add configuration to support
special periodic or one-shot handling for each discovered
modbus device belonging to the specific register map.

This is primarily targeted towards adding support for periodic
updating of timestamp register of PSUs.

Test Plan:
Schema is validated against future configuration added
which in turn is validated on a physical device for
functional success.

Reviewed By: garnermic

fbshipit-source-id: e47686bbcfc1d42dfcc934e9782ea5afbcd040f3