rackmon2: disable special handler if not supported
Summary:
If the PSU does not support writing to the given register
(Either writing is not supported or the specific register
is not available), then disable the handler to prevent
future updates which is just adding noise to error logs.

Test Plan:
load rackmon on a system where a4,a5,a6,b4,b5 are gen2 and support
writing to the timestamp register. Only B6 is gen1 which does
not support. Ensure we get the correct error message
```
  2022 Jun 29 16:32:41 $HOST user.info wedge100-2022.11.1: rackmond[17149]: DEV:0xb6 Special Handler at 0x12a  unsupported. Disabled from future updates
```

Reviewed By: garnermic

fbshipit-source-id: becdcd0e2ec8ebb43302dbbb765219b1380fd0ed