meta-aspeed: Add SRCPV to kernel and u-boot recipes
Summary:
I think because of
[this](https://lore.kernel.org/all/CANNYZj8mR7Ov9w+=b3mv1QVLc_0G9Nusj1D3ZmzYq3yn3pA5MA@mail.gmail.com/T/)
commit in bitbake, we now need to include SRCPV in AUTOREV recipes.

I also incremented the kernel version from .25 to .27, because the
revision has increased. I think it's expected that we need to manually
increment the kernel version every so often with AUTOREV.

Test Plan: Watching netcastle build jobs for kirkstone platforms.

Reviewers: goldenbug

Subscribers:

Tasks:

Tags:

Differential Revision: https://phabricator.intern.facebook.com/D37801554