common:ipmid-v2: platform-specific configuration file
Summary:
this is the platform-specific configuration for elbert. it can be used as the template for other new NPIs.

The ipmid-v2 is a redesign of ipmid with the goal of making the new NPI introduction easier and reducing the code maintenance cost.
The ipmid-v2 is changed to be platform-specific configuration driving without PALLIB support.

Test Plan:
The ipmid-v2 is tested by running ipmitool on the X86 side.  ipmitool will communicate with ipmid-v2 in BMC side.
tested various ipmitool subcommands:
ipmitool  mc info
ipmitool fru
ipmitool  mc guid
ipmitool raw
tracing the ipmi messages on the stack: ipmid/kcsd on the BMC side.

Reviewed By: tao-ren

fbshipit-source-id: 82d7e30bda6488e9b75089239e3d9fdfe489f07b