attest-util: Allow users to pass in device I2C bus and addr
Summary:
As title, instead of hard-coding it, let users pass it in.
TODO: This might be better in the future to have a configuration
file to map endpoint ID to bus:addr per platform.

This unblocks us to test it out on newer platforms till we have
that functionality.

Test Plan: Build on all platforms.

Reviewed By: williamspatrick

fbshipit-source-id: 046da5f0c154673bd0090c2c1b34a2c9a24bfd9e