recipes-devtools: Add qemu-system-native_git.bb
Summary:
The upstream QEMU recipe `qemu-system-native_7.0.0.bb` bases off of release tarballs, this recipe modifies it slightly to support tracking the git repository directly.

`gitsm://` didn't seem to work, so I specified the minimum set of submodules as `git://` URI's, as suggested by Patrick and [this thread](https://www.openembedded.org/pipermail/bitbake-devel/2019-August/020301.html).

I removed the patch files because I couldn't figure out how to change `FILESEXTRAPATHS` to find them, and I also don't think we really care about them.

This is easy to backport to kirkstone, not terrible to put in dunfell, but basically impossible to get in rocko. But the `qemu-system-native` recipe is not included in the image dependencies on rocko.

Test Plan: Making sure that all platforms' build jobs succeed.

Reviewed By: williamspatrick

fbshipit-source-id: 2e4c23234a1f49bd893b08d31f20cd725dcda794