rackmon2: Add debug trace capability
Summary:
Add support to (default off) optionally print a trace of
logs received/transmitted message over any of the supported
interfaces.

This is implemented as a interface configuration so we can use it to
debug specific issues on an interface rather than globally, and also
not require special debug builds.

Test Plan:
Change /etc/rackmon.conf to add `"debug": true` to the /dev/ttyUSB1
interface for a wedge400 respin BMC:
```
 2022 Aug  9 15:03:31 <host> user.info wedge400-2022.27.0: rackmond[14954]: /dev/ttyUSB1 TX: 0x6503000800010dec
 2022 Aug  9 15:03:31 <host> user.info wedge400-2022.27.0: rackmond[14954]: /dev/ttyUSB1 RX: 0x6503023730dfa8
 2022 Aug  9 15:03:31 <host> user.info wedge400-2022.27.0: rackmond[14954]: Found 65 on /dev/ttyUSB1
```
Ensure no prints when the configuration is removed and rackmon is
restarted.

Reviewed By: garnermic

fbshipit-source-id: bf66da4dc136cd957f89746fe189427446b15aff