qemu: Enable runqemu for all supported machines
Summary:
This enables qemuboot configuration generation, which is used by runqemu
to figure out how to boot an image in QEMU. With this commit, you can
build an image for any of these machines and use `runqemu` to boot it
with all the right arguments.

The configuration I've added, in particular this:

```
EXTRA_IMAGEDEPENDS:append = " qemu-helper-native:do_addto_recipe_sysroot"
```

Only seems to work on kirkstone and newer yocto branches, dunfell
doesn't like the colon.

Additionally, it seems to *require* the `do_addto_recipe_sysroot` task
explicitness because the runqemu script looks in a weird directory for
QEMU, and without that, it doesn't get populated.

Test Plan:
On machines that it does work with, you can do this:

```
source openbmc-init-build-env <machine> build
bitbake <machine>-image
runqemu nographic
runqemu nographic slirp
```

Reviewed By: williamspatrick

fbshipit-source-id: 620363784a1938b4c1b9a2ba35fe077e912f776e