kernel: fix PV for kirkstone+ distros
Summary:
Kirkstone and later distros require the "floating branch"
packages to specify the PV in terms of the SRCREV.  Adding the
identifer `+git${SRCPV}` is the typical method in Yocto recipes to
achieve this.  Update all the kernel github recipes to accomplish
this.

Test Plan: https://github.com/facebook/openbmc/pull/201

Reviewed By: peterdelevoryas

fbshipit-source-id: 4bbccbfe2222398df955c28db02c2530f5147533