u-boot: fix PV for kirkstone+ distros
Summary:
Kirkstone and later distros require the "floating branch"
packages to specify the PV in terms of the SRCREV.  Adding the
identifer `+git${SRCPV}` is the typical method in Yocto recipes to
achieve this.  Update all the u-boot recipes to accomplish this.

For internal build, the u-boot%.bbappend will modify the PV and SRC_URI back to use in-tree local files.

The u-boot%.bbappend will be applied to all u-boot related bb recipes,  while we only want to apply it to recipes using our u-boot-common.inc so add uboot-aspeed-fb as conditional override to skip modify PV and SRC_URI of u-boot recipes from yocto.

Test Plan: https://github.com/facebook/openbmc/pull/201

Reviewed By: peterdelevoryas

fbshipit-source-id: 28d9ec7404606f846e0e07976d376412868ad0f2