fbqemu: Move to common/recipes-devtools
Summary:
Previously, the fbqemu recipe only even existed in kirkstone and
lf-master, by placing the recipe in the backports directories.

But, to add fbqemu-native to `lf-obmc download-mirror`, it would make
things a lot simpler if we add it unconditionally for every machine and
branch. That way we can keep the weird backports logic in
meta-backports, and not in lf_obmc.

Test Plan:
Verified I can build fbqemu-native with a machine from rocko, dunfell,
kirkstone, and lf-master.

Reviewed By: williamspatrick

fbshipit-source-id: ee0944d0f2858af1e54541e68089d9668a69ad1f