recipes-devtools: Delete fbqemu
Summary:
I moved this tool to `fbsource`, so it's not needed here anymore.

If we make it available to ODM's, it would be by mirroring the fbsource
directory and publishing to crates.io, installing with `cargo install
bmcqemu`.

This also eliminates our dependency on rust in image building, seeing as it's really expensive (large) and not part of the image itself.

Test Plan: N/A

Reviewed By: williamspatrick

fbshipit-source-id: 7e0995b0271bf6b32095998f6cd58fb285669ec8