greatlakes: Add jtag, pwm, wdt, tpm patches
Summary:
This is all needed for booting on hardware. Wiwynn validated these, and they posted the patches to the LF codebase [here](https://gerrit.openbmc.org/c/openbmc/openbmc/+/56429/1), but it's not allowed upstream.

1. Adds a jtag driver
2. Adds a PWM driver
3. Adds a fix for the watchdog
4. Updates the aspeed-g6.dtsi
5. Updates the greatlakes dts to enable the jtag, pwm's, and tpm.

The kernel config changes are already in the greatlakes.cfg in LF OpenBMC.

Test Plan: Have Wiwynn test this image again.

Reviewed By: GoldenBug

fbshipit-source-id: a051329fd147775b07d2d1928dae5f8d481cb55d