sandcastle: Add test for job.py
Summary:
This adds a test for `facebook/sandcastle/job.py`. I think the current behavior is actually pretty unexpected, in particular we're scheduling too many QEMU jobs, and not scheduling CIT for any Compute/Storage platforms on-diff.

Check this P529607482 to see what I mean. It's a dump of (purpose, platform, netcastle command) pairs.

Test Plan:
```
$ ./facebook/sandcastle/test.py
..
----------------------------------------------------------------------
Ran 2 tests in 0.008s

OK
```

Verified the Sandcastle job produces an error in a do-not-merge diff D39517026:

{F770130744}

Reviewed By: GoldenBug

fbshipit-source-id: aab0d478e34e749a26c9921c7f17512eb57f1b96