backport DHCPv6.SendVendorOption to dunfell
Summary:
We need to send model name and serial number with DHCPv6 requests from the
BMC.  We contributed code to systemd in 2020 that allows sending vendor
options along with DHCPv6 requests.

We need this code to move wedge100 over to Yocto Dunfell.  The systemd
version in Dunfell is old and doesn't have this code.  Therefore, backport the
change to unblock the update of wedge100.

Reviewed By: kawmarco

fbshipit-source-id: c8487de400791fa7be9103b2d63786c36061ec16