utd: Fix openbmc repository-state
Summary: We need to make sure that we specify the state for fbsource *and* openbmc in our UTD rule. We specifically need the everstore hash added by UTD on the OpenBMC repository state, and that's done automatically if we use `repo_name="openbmc"`. The `skycastle_keyword_options` repo states won't get an everstore hash added to them, but that's ok, cause we just need the public stable hash for fbsource.

Test Plan:
Verifying the git hash scheduled in Signal Box skycastle jobs.
```
$ git rev-parse HEAD
390a73cc348c750164f68b9fe6a618cd36f0b56a
```
https://fburl.com/skycastle/iiwggbzq

{F775269848}

fbshipit-source-id: b8329395a4b7ee4f21d27bb37463c92c9679859c